pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts>=0.2.1",
    "discord-py>=2.4.0",
    "mistralai>=1.4.0",
    "python-dotenv>=1.0.1",
    "yfinance>=0.2.36",
    "alpha-vantage>=2.3.1",
    "pandas>=2.2.0"
]
